+++ blockchain/src/SampleNFTBuyer.sol
@@ -13,16 +13,22 @@ contract SampleNFTBuyer is ERC721URIStorage, Ownable, Sample {
         tokenCounter = 1;
     }
 
-	modifier onlyCanPay(address _scNFT, uint _idNFT) {
-		require(msg.value >= Sample.addressToTokenIdToPrice[_scNFT][_idNFT], "Not enough money");
-		_;
-	}
+    modifier onlyCanPay(address _scNFT, uint256 _idNFT) {
+        require(msg.value >= Sample.addressToTokenIdToPrice[_scNFT][_idNFT], "Not enough money");
+        _;
+    }
 
-    function mintNFT(address _scNFT, uint _idNFT, string memory tokenURI) public payable onlyOwner onlyCanPay(_scNFT, _idNFT) returns (uint256) {
+    function mintNFT(address _scNFT, uint256 _idNFT, string memory tokenURI)
+        public
+        payable
+        onlyOwner
+        onlyCanPay(_scNFT, _idNFT)
+        returns (uint256)
+    {
         uint256 newTokenId = tokenCounter++;
         _safeMint(msg.sender, newTokenId);
         _setTokenURI(newTokenId, tokenURI);
-		payable(Sample.ownerToNFT[_scNFT]).transfer(Sample.addressToTokenIdToPrice[_scNFT][_idNFT]);
+        payable(Sample.ownerToNFT[_scNFT]).transfer(Sample.addressToTokenIdToPrice[_scNFT][_idNFT]);
         return newTokenId;
     }
 }
